% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regsim.R
\name{plotRegsim}
\alias{plotRegsim}
\title{Plotting Parameters}
\usage{
plotRegsim(reps = 100, ns = c(10, 25, 100), betaall = seq(-1, 1, 0.1),
  err.dist = "t", alph = 0.05, ...)
}
\arguments{
\item{reps}{positive integer of number of repetitions of simulation}

\item{ns}{list of positive integers of sample sizes}

\item{betaall}{list of different \eqn{\beta} values to test}

\item{err.dist}{a string indicating what distribution to use for errors. The current build
supports \code{uniform}, \code{t}, \code{laplace}}

\item{alph}{a double between 0 and 1, representing the siginificance level \eqn{\alpha}
to test. Default is 0.05.}

\item{...}{additional parameters for the err.dist. See details for more information.}
}
\value{
\code{plotRegsim} returns a list of 6 ggplot objects. The first three are plots related to
having a normally distributed error, while the remaining are plots related to the error
distribution provided by the user. Each plot colors the data based on sample size. Specifically

\enumerate{

\item Power vs. beta, normal errors

\item Beta.hat vs. beta, normal errors

\item r^2 vs. beta, normal errors

\item Power vs. beta, uniform/t/laplace errors

\item Beta.hat vs. beta, uniform/t/laplace errors

\item r^2 vs. beta, uniform/t/laplace errors

}
}
\description{
Visualizing the relation between different parameters and estimates is useful. This function
plots estimates of slope, power, and r^2 from a simple linear regression model that is
checked for assumption violation via \link{regsim}. These plots are
created by \code{ggplot}.
}
\examples{
plotRegsim(err.dist = "laplace", rate=3)

}
\seealso{
\link[project]{regsim}

\link[ggplot2]{ggplot}
}
