% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regsim.R
\name{regsim}
\alias{regsim}
\title{Assumption Violation Simulation}
\usage{
regsim(reps = 100, ns = 100, betaall = seq(-1, 1, 0.1), alph = 0.05,
  err.dist = "t", ...)
}
\arguments{
\item{reps}{positive integer of number of repetitions of simulation}

\item{ns}{list of positive integers of sample sizes}

\item{betaall}{list of different \eqn{\beta} values to test}

\item{alph}{a double between 0 and 1, representing the siginificance level \eqn{\alpha}
to test. Default is 0.05.}

\item{err.dist}{a string indicating what distribution to use for errors. The current build
supports \code{uniform}, \code{t}, \code{laplace}}

\item{...}{additional parameters for the err.dist. See details for more information.}
}
\value{
\code{regsim()} returns a list that contains two lists \code{norm.err} and
\code{oth.err}. Each of these lists contains two data.frames. Both contain the
two columns of sample sizes \code{n} and parameter \code{beta}.

The first data.frame \code{pow} contains a 3rd column of
\code{power}, estimates of power. This data.frame is intended for examining power of the
test. In summary, \code{pow} contains
\itemize{
\item \code{n}
\item \code{beta}
\item \code{power} Estimate of power
}
The second data.frame \code{est} contains a 2 additional columns for estimates
of beta, \code{beta.hat}, and the coefficient of determination, r^2 values.
In summary, \code{est} contains
\itemize{
\item \code{n}
\item \code{beta}
\item \code{beta.hat} Estimate of beta
\item \code{r.sq} R^2 values
}
}
\description{
One of the assumptions for linear regression is normally distributed errors. If
this is not met, the standard errors can be invalidated. This function runs a
simulation with a normal distribution and with non-normal error distribution,
and returing the distributions of various estimates of parameters, such as the
power of the test.
}
\details{
It is important to check assumptions before performing linear regression. This function
runs a simulation twice, once with normal errors and once with non-normal errors. The
current build supports \code{uniform}, \code{t}, and \code{laplace} as non-normal
distributions. The default is \code{t} with degrees of freedom 10. Since there is
no built in laplace function, we generate values from a difference of two exponential
distributions, which is equivalent to a laplace distirbution.

To estimate \code{beta}, we use \eqn{beta.hat = cov(x,y)/var(x)}.

To calculate r^2, we use \deqn{1 - (SSR / SST)}
where SSE is the residuals sum of squares and SST is the total sum of squares.
SSR is extracted from the \code{\link[stats]{lm.fit}}{.lm.fit()} function.

Hypothesis testing with simple linear regression is equivalent to anova. However, since
running \code{anova} requires the argument of a model (also using \code{lm}), it is faster to
calculate the p-value by using \code{pf} and using the F-statistic calculated as
\deqn{r^2 / (1-r^2) * (n-1)} This is derived from the relation between r^2 and the
F-statistic \deqn{r^2/(1-r^2)*df_2/df_1} where df_1 is the degrees of freedom for the
number of predictors, and df_2 is the degrees of freedom for the full model.

The three distributions that the user can check for assumption violation are
\code{t}, \code{uniform}, \code{laplace}.
\itemize{
\item For \code{t}, the additional argument is \code{df}, which must be positive.
\item For \code{uniform}, the additional arguments are \code{min} and \code{max}.
\item For \code{laplace}, the additional argument is the \code{rate}, which must be positive.
}
See examples for syntax.
}
\examples{

regsim(err.dist = "uniform", min=-1,max=1)
regsim(err.dist = "t", df=25)
regsim(err.dist = "laplace", rate=3)

}
\references{
\url{http://dept.stat.lsa.umich.edu/~jerrick/courses/stat701/index.html}

\url{http://r-pkgs.had.co.nz/}

\url{https://en.wikipedia.org/wiki/Laplace_distribution#Generating_random_variables_according_to_the_Laplace_distribution}

\url{https://stats.stackexchange.com/questions/56881/whats-the-relationship-between-r2-and-f-test}

\url{https://en.wikipedia.org/wiki/F-test}

\url{http://stackoverflow.com/questions/32585753/r-swap-two-variables-without-using-a-third}

\url{http://www.deanbodenham.com/learn/r-package-documentation.html}
}
\seealso{
\link[stats]{lm.fit}

\link{plotRegsim}
}
